backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
pymongo==4.3.3
pandas==1.5.3
scikit-learn==1.2.2
textblob==0.17.1
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.1
python-dotenv==1.0.0 